refactor(desktop): make Nucleus the sole desktop app
