--- desktopApp/native/linux-tray/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "fuoevolve-linux-tray-bridge"
-version = "0.1.0"
-edition = "2021"
-rust-version = "1.80"
-publish = false
-
-[lib]
-name = "fuoevolve_linux_tray_bridge"
-crate-type = ["cdylib"]
-
-[target.'cfg(target_os = "linux")'.dependencies]
-png = "=0.18.1"
-ksni = { version = "=0.3.6", features = ["blocking"] }
