--- desktopApp/native/macos-now-playing/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "fuoevolve-now-playing-bridge"
-version = "0.1.0"
-edition = "2021"
-rust-version = "1.76"
-publish = false
-
-[lib]
-name = "fuoevolve_now_playing_bridge"
-crate-type = ["cdylib"]
-
-# playwire owns the MPNowPlayingInfoCenter / MPRemoteCommandCenter Objective-C bridge.
-# Keep the JVM boundary as C ABI + JNA so Apple framework types never leak into shared code.
-[target.'cfg(target_os = "macos")'.dependencies]
-playwire = "=1.0.0"
-objc2 = "=0.6.4"
-objc2-foundation = { version = "=0.3.2", features = ["NSDictionary", "NSObject", "NSString", "NSValue"] }
-objc2-media-player = { version = "=0.3.2", features = ["MPNowPlayingInfoCenter"] }
