--- desktopApp/native/windows-smtc/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "fuoevolve-smtc-bridge"
-version = "0.1.0"
-edition = "2021"
-rust-version = "1.95"
-publish = false
-
-[lib]
-name = "fuoevolve_smtc_bridge"
-crate-type = ["cdylib"]
-
-[target.'cfg(windows)'.dependencies]
-windows = { version = "0.62.2", features = [
-    "Foundation",
-    "Media",
-    "Storage_Streams",
-    "Win32_Foundation",
-    "Win32_System_WinRT",
-] }
