+++ desktopApp/packaging/linux/prepare-portable-runtime.sh
