+++ desktopApp/packaging/windows/prepare-libmpv-dev.ps1
