+++ desktopApp/src/main/resources/META-INF/native-image/org.feeluown/fuoevolve/reachability-metadata.json
