+++ desktopApp/src/test/kotlin/org/feeluown/mobile/nucleus/NucleusExternalActivationTest.kt
